+++ CumulusMX/Utils.cs
@@ -192,42 +192,6 @@ namespace CumulusMX
 		}
 
 
-		public static DateTime AmbiguousDateToLocal(DateTime d, ref List<DateTime> dates)
-		{
-			// if ambiguous attempt to map to correct DST offset
-			if (TimeZoneInfo.Local.IsAmbiguousTime(d))
-			{
-				// This localDt maps to two possible UTC instants (fall-back overlap).
-				// You can choose:
-				//  - the earlier offset (daylight) or
-				//  - the later offset (standard)
-				var offsets = TimeZoneInfo.Local.GetAmbiguousTimeOffsets(d);
-				TimeSpan chosenOffset;
-				if (dates.Count > 0 && dates[^1] >= d)
-				{
-					chosenOffset = offsets[0];
-				}
-				else
-				{
-					chosenOffset = offsets[1];
-					dates.Add(d);
-				}
-
-				var dto = new DateTimeOffset(d, chosenOffset);
-				return dto.LocalDateTime;
-			}
-			else
-			{
-				// A “normal” unambiguous time
-				dates.Clear();
-				var offset = TimeZoneInfo.Local.GetUtcOffset(d);
-				var dto = new DateTimeOffset(d, offset);
-				return dto.LocalDateTime;
-			}
-		}
-
-
-
 		public static IPAddress GetIpWithDefaultGateway()
 		{
 			try